+++ R/dengue_SP.R
@@ -2,9 +2,9 @@
 #'
 #' Data frame containing the weekly number of notified dengue cases in
 #' the municipality of *Sao Paulo*, as well as a set of climatic covariates.
-#' 
+#'
 #' @docType data
-#' @format 
+#' @format
 #' A data frame with 678 rows and 8 columns:
 #' \describe{
 #'   \item{date}{First day of the week in date format.}