+++ tests/testthat/test-plot_correlation.R
@@ -1,11 +1,16 @@
 test_that("error if data is not a data.frame", {
-  expect_error(plot_correlation(data = list(a = 1), var = "a"), 
-               "'data' should be a 'data.frame'")
+  expect_error(
+    plot_correlation(data = list(a = 1), var = "a"),
+    "'data' should be a 'data.frame'"
+  )
 })
 
 test_that("error if variables are not numeric", {
   df <- data.frame(num1 = rnorm(10), cat1 = letters[1:10])
-  expect_error(plot_correlation(df, var = c("num1", "cat1")), "variables should be numeric")
+  expect_error(
+    plot_correlation(df, var = c("num1", "cat1")),
+    "variables should be numeric"
+  )
 })
 
 test_that("returns a ggplot object for valid inputs", {
@@ -16,7 +21,11 @@ test_that("returns a ggplot object for valid inputs", {
 
 test_that("uses var_label when provided", {
   df <- data.frame(a = rnorm(10), b = rnorm(10))
-  expect_no_error(plot_correlation(df, var = c("a", "b"), var_label = c("Var A", "Var B")))
+  expect_no_error(plot_correlation(
+    df,
+    var = c("a", "b"),
+    var_label = c("Var A", "Var B")
+  ))
 })
 
 test_that("handles custom title correctly", {
@@ -32,9 +41,21 @@ test_that("handles palette argument", {
 
 test_that("works with different correlation methods", {
   df <- data.frame(a = rnorm(10), b = rnorm(10), c = rnorm(10))
-  expect_no_error(plot_correlation(df, var = c("a", "b", "c"), method = "pearson"))
-  expect_no_error(plot_correlation(df, var = c("a", "b", "c"), method = "spearman"))
-  expect_no_error(plot_correlation(df, var = c("a", "b", "c"), method = "kendall"))
+  expect_no_error(plot_correlation(
+    df,
+    var = c("a", "b", "c"),
+    method = "pearson"
+  ))
+  expect_no_error(plot_correlation(
+    df,
+    var = c("a", "b", "c"),
+    method = "spearman"
+  ))
+  expect_no_error(plot_correlation(
+    df,
+    var = c("a", "b", "c"),
+    method = "kendall"
+  ))
 })
 
 test_that("prints correlation matrix if print = TRUE", {
@@ -51,23 +72,35 @@ test_that("handles NA values in data", {
 
 test_that("error for invalid plot_type", {
   df <- data.frame(num1 = rnorm(10), num2 = rnorm(10))
-  expect_error(plot_correlation(df, var = c("num1", "num2"), plot_type = 1),
-               "plot_type must be a character vector of length 2")
-  expect_error(plot_correlation(df, var = c("num1", "num2"), plot_type = c(1,"circle")),
-               "Accepted plot_type include: 'raster', 'number', 'circle'")
+  expect_error(
+    plot_correlation(df, var = c("num1", "num2"), plot_type = 1),
+    "plot_type must be a character vector of length 2"
+  )
+  expect_error(
+    plot_correlation(df, var = c("num1", "num2"), plot_type = c(1, "circle")),
+    "Accepted plot_type include: 'raster', 'number', 'circle'"
+  )
 })
 
 
 test_that("error for invalid plot_type", {
   df <- data.frame(num1 = rnorm(10), num2 = rnorm(10))
-  result <- plot_correlation(df, var = c("num1", "num2"), 
-                             plot_type = c("raster", "circle"))
+  result <- plot_correlation(
+    df,
+    var = c("num1", "num2"),
+    plot_type = c("raster", "circle")
+  )
   expect_s3_class(result, "ggplot")
-  result <- plot_correlation(df, var = c("num1", "num2"), 
-                             plot_type = c("raster", "number"))
+  result <- plot_correlation(
+    df,
+    var = c("num1", "num2"),
+    plot_type = c("raster", "number")
+  )
   expect_s3_class(result, "ggplot")
-  result <- plot_correlation(df, var = c("num1", "num2"), 
-                             plot_type = c("number", "circle"))
+  result <- plot_correlation(
+    df,
+    var = c("num1", "num2"),
+    plot_type = c("number", "circle")
+  )
   expect_s3_class(result, "ggplot")
 })
-