+++ tests/testthat/test-plot_multiple.R
@@ -6,11 +6,17 @@ test_that("error if plot_function is not one of the allowed functions", {
 
 test_that("returns list of ggplots from plot_timeseries", {
   df <- data.frame(
-    date = rep(seq.Date(from = as.Date("2020-01-01"), by = "month", length.out = 6), 2),
+    date = rep(
+      seq.Date(from = as.Date("2020-01-01"), by = "month", length.out = 6),
+      2
+    ),
     region = c(rep("A", 6), rep("B", 6)),
-    x = rnorm(12), y = rnorm(12), z = rnorm(12), pop = rnorm(12)/100
+    x = rnorm(12),
+    y = rnorm(12),
+    z = rnorm(12),
+    pop = rnorm(12) / 100
   )
-  
+
   plots <- plot_multiple(
     plot_function = plot_timeseries,
     data = df,
@@ -22,7 +28,7 @@ test_that("returns list of ggplots from plot_timeseries", {
     time = "date",
     area = "region"
   )
-  
+
   expect_type(plots, "list")
   expect_length(plots, 3)
   expect_s3_class(plots[[1]], "ggplot")
@@ -32,11 +38,17 @@ test_that("returns list of ggplots from plot_timeseries", {
 
 test_that("returns list of ggplots from plot_heatmap", {
   df <- data.frame(
-    date = rep(seq.Date(from = as.Date("2020-01-01"), by = "month", length.out = 6), 2),
+    date = rep(
+      seq.Date(from = as.Date("2020-01-01"), by = "month", length.out = 6),
+      2
+    ),
     region = c(rep("A", 6), rep("B", 6)),
-    x = rnorm(12), y = rnorm(12), z = rnorm(12), pop = rnorm(12)/100
+    x = rnorm(12),
+    y = rnorm(12),
+    z = rnorm(12),
+    pop = rnorm(12) / 100
   )
-  
+
   plots <- plot_multiple(
     plot_function = plot_heatmap,
     data = df,
@@ -48,7 +60,7 @@ test_that("returns list of ggplots from plot_heatmap", {
     time = "date",
     area = "region"
   )
-  
+
   expect_type(plots, "list")
   expect_length(plots, 3)
   expect_s3_class(plots[[1]], "ggplot")
@@ -59,11 +71,17 @@ test_that("returns list of ggplots from plot_heatmap", {
 
 test_that("returns list of ggplots from plot_seasonality", {
   df <- data.frame(
-    date = rep(seq.Date(from = as.Date("2020-01-01"), by = "month", length.out = 6), 2),
+    date = rep(
+      seq.Date(from = as.Date("2020-01-01"), by = "month", length.out = 6),
+      2
+    ),
     region = c(rep("A", 6), rep("B", 6)),
-    x = rnorm(12), y = rnorm(12), z = rnorm(12), pop = rnorm(12)/100
+    x = rnorm(12),
+    y = rnorm(12),
+    z = rnorm(12),
+    pop = rnorm(12) / 100
   )
-  
+
   plots <- plot_multiple(
     plot_function = plot_seasonality,
     data = df,
@@ -75,7 +93,7 @@ test_that("returns list of ggplots from plot_seasonality", {
     time = "date",
     area = "region",
   )
-  
+
   expect_type(plots, "list")
   expect_length(plots, 3)
   expect_s3_class(plots[[1]], "ggplot")
@@ -86,14 +104,20 @@ test_that("returns list of ggplots from plot_seasonality", {
 
 test_that("returns list of ggplots from plot_map", {
   data("map_MS")
-  map_df <- map_MS[1:2,]
+  map_df <- map_MS[1:2, ]
   map_df$code <- c("A", "B")
   df <- data.frame(
-    date = rep(seq.Date(from = as.Date("2020-01-01"), by = "month", length.out = 6), 2),
+    date = rep(
+      seq.Date(from = as.Date("2020-01-01"), by = "month", length.out = 6),
+      2
+    ),
     region = c(rep("A", 6), rep("B", 6)),
-    x = rnorm(12), y = rnorm(12), z = rnorm(12), pop = rnorm(12)/100
+    x = rnorm(12),
+    y = rnorm(12),
+    z = rnorm(12),
+    pop = rnorm(12) / 100
   )
-  
+
   plots <- plot_multiple(
     plot_function = plot_map,
     data = df,
@@ -107,7 +131,7 @@ test_that("returns list of ggplots from plot_map", {
     area = "region",
     map_area = "code",
   )
-  
+
   expect_type(plots, "list")
   expect_length(plots, 3)
   expect_s3_class(plots[[1]], "ggplot")